# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TravelCar API",
        "description": "API for managing travel-related data, including user authentication, trips, etc.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/trips": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Get all trips",
                "operationId": "584f2b36d970bf78a2860728d5559cb8",
                "responses": {
                    "200": {
                        "description": "List of trips",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Trips"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Trips"
                ],
                "summary": "Create a new trip",
                "operationId": "37095c0de2427dc1f8b2a4a1f0be2d95",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "starting_point",
                                    "ending_point",
                                    "starting_at",
                                    "available_places",
                                    "price"
                                ],
                                "properties": {
                                    "starting_point": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "ending_point": {
                                        "type": "string",
                                        "example": "London"
                                    },
                                    "starting_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-07-23T14:30:00Z"
                                    },
                                    "available_places": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Trip created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trips"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/trips/{id}": {
            "get": {
                "tags": [
                    "Trips"
                ],
                "summary": "Get a trip by ID",
                "operationId": "e0b6d44323e5fbc7d7c4ed0d9089e77c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trip details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trips"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trip not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Trips"
                ],
                "summary": "Update a trip",
                "operationId": "e9c1fdc21866df3d46c108b60e7b37e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "starting_point": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "ending_point": {
                                        "type": "string",
                                        "example": "London"
                                    },
                                    "starting_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-07-23T14:30:00Z"
                                    },
                                    "available_places": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Trip updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Trips"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Trip not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Trips"
                ],
                "summary": "Delete a trip",
                "operationId": "1c635a44aca5befbf430ae4abe25f52c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Trip deleted successfully"
                    },
                    "404": {
                        "description": "Trip not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Trips": {
                "required": [
                    "starting_point",
                    "ending_point",
                    "starting_at",
                    "available_places",
                    "price",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the trip",
                        "type": "integer"
                    },
                    "starting_point": {
                        "description": "Starting point of the trip",
                        "type": "string"
                    },
                    "ending_point": {
                        "description": "Ending point of the trip",
                        "type": "string"
                    },
                    "starting_at": {
                        "description": "Starting date and time of the trip",
                        "type": "string",
                        "format": "date-time"
                    },
                    "available_places": {
                        "description": "Number of available places",
                        "type": "integer"
                    },
                    "price": {
                        "description": "Price of the trip",
                        "type": "number",
                        "format": "float"
                    },
                    "user_id": {
                        "description": "ID of the user who created the trip",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}